+++ app/javascript/vue/components/Filter/Facets/Otu/components/Includes.vue
@@ -36,8 +36,8 @@ const includes = [
   'collection_objects',
   'collection_object_observations',
   'otu_observations',
-  'type_materials',
-  'type_material_observation'
+  'type_material',
+  'type_material_observations'
 ]
 
 const params = computed({