+++ app/javascript/vue/tasks/collection_objects/filter/components/filter.vue
@@ -28,9 +28,9 @@
   <FacetLoan v-model="params" />
   <FacetUsers v-model="params" />
   <FacetTags
-     v-model="params"
-     target="CollectionObject"
-   />
+    v-model="params"
+    target="CollectionObject"
+  />
   <FacetNotes v-model="params" />
   <FacetBuffered v-model="params" />
   <FacetProtocol v-model="params" />
@@ -69,6 +69,7 @@ import FacetDataAttribute from 'components/Filter/Facets/shared/FacetDataAttribu
 import FacetWKT from 'components/Filter/Facets/Otu/FacetWKT.vue'
 import FacetMatchIdentifiers from 'components/Filter/Facets/shared/FacetMatchIdentifiers.vue'
 import FacetProtocol from 'components/Filter/Facets/Extract/FacetProtocol.vue'
+import FacetImportDataset from 'components/Filter/Facets/CollectionObject/FacetImportDataset.vue'
 
 const WITH_PARAMS = [
   'biological_associations',
@@ -77,24 +78,27 @@ const WITH_PARAMS = [
   'collectors',
   'current_repository',
   'data_attributes',
+  'data_depictions',
   'dates',
   'deaccessioned',
+  'depictions',
   'determiners',
   'dwc_indexed',
   'geographic_area',
   'georeferences',
-  'depictions',
-  'local_identifiers',
   'global_identifiers',
+  'local_identifiers',
   'notes',
   'origin_citation',
   'preparation_type',
+  'protocols',
   'repository',
+  'tags',
   'taxon_determinations',
   'type_material',
   'with_buffered_collecting_event',
   'with_buffered_determinations',
-  'with_buffered_other_labels'
+  'with_buffered_other_labels',
 ]
 
 const emit = defineEmits(['update:modelValue'])