+++ lib/queries/geographic_area/autocomplete.rb
@@ -25,20 +25,42 @@ module Queries
       def autocomplete
         return [] if query_string.blank?
         queries = [
-          ::GeographicArea.where(id: query_string).all,
-          ::GeographicArea.where(name: query_string).all,
-          ::GeographicArea.joins(parent_child_join).where(Arel.sql(parent_child_where.to_sql)).limit(5).all,
-          ::GeographicArea.where(Arel.sql(where_sql)).includes(:geographic_area_type, :geographic_items).order(Arel.sql('LENGTH(name)')).limit(dynamic_limit).all,
-           autocomplete_exact_id,
-           autocomplete_identifier_cached_exact,
-           autocomplete_identifier_identifier_exact,
-           autocomplete_alternate_values_name.limit(20)
+          [ ::GeographicArea.where(id: query_string).all, false ],
+          [ ::GeographicArea.where(name: query_string).all, true ],
+          [ ::GeographicArea.joins(:alternate_values).where('alternate_values.value ILIKE ?', query_string).all, true ],
+          [ ::GeographicArea.joins(parent_child_join).where(Arel.sql(parent_child_where.to_sql)).limit(5).all, true ],
+          [ ::GeographicArea.where(Arel.sql(where_sql)).limit(dynamic_limit).all, true ],
+          [ autocomplete_exact_id, false ],
+          [ autocomplete_identifier_cached_exact, false ],
+          [ autocomplete_identifier_identifier_exact, false ],
+          [ autocomplete_alternate_values_name.limit(20), true ]
         ]
 
         queries.compact!
 
+        updated_queries = []
+        pr_id = project_id.join(',') if project_id
+
+        queries.each_with_index do |q, i|
+          a = q[0]
+
+          if q[1] # do not use extended query for identifiers
+            if project_id.present?
+
+              if !a.nil?
+                a = a.left_outer_joins(:asserted_distributions)
+                     .left_outer_joins(:collecting_events)
+                     .select("geographic_areas.*, (COUNT(collecting_events.id) + COUNT(asserted_distributions.id)) AS use_count, CASE WHEN asserted_distributions.project_id IN (#{pr_id}) THEN asserted_distributions.project_id WHEN collecting_events.project_id IN (#{pr_id}) THEN collecting_events.project_id ELSE NULL END AS in_project")
+                     .group('geographic_areas.id, collecting_events.project_id, asserted_distributions.project_id')
+                     .order('in_project, use_count DESC')
+              end
+            end
+          end
+          updated_queries[i] = a
+        end
         result = []
-        queries.each do |q|
+        updated_queries.compact!
+        updated_queries.each do |q|
           result += q.includes(:geographic_areas_geographic_items).to_a
           result.uniq!
           break if result.count > 19